+++ rmk/src/config/mod.rs
@@ -10,6 +10,7 @@ use macro_config::KeyboardMacrosConfig;
 
 use crate::combo::Combo;
 use crate::fork::Fork;
+use crate::morse::MorseKeyMode;
 use crate::tap_dance::TapDance;
 use crate::{COMBO_MAX_NUM, FORK_MAX_NUM, TAP_DANCE_MAX_NUM};
 
@@ -28,7 +29,7 @@ pub struct RmkConfig<'a> {
 #[derive(Debug, Default)]
 pub struct BehaviorConfig {
     pub tri_layer: Option<[u8; 3]>,
-    pub tap_hold: TapHoldConfig,
+    pub morse: MorseConfig,
     pub one_shot: OneShotConfig,
     pub combo: CombosConfig,
     pub fork: ForksConfig,
@@ -49,36 +50,27 @@ impl Default for TapDancesConfig {
     }
 }
 
-/// Mode for tap hold behavior
+/// Configurations for morse behavior
 #[derive(Clone, Copy, Debug)]
-#[cfg_attr(feature = "defmt", derive(defmt::Format))]
-pub enum TapHoldMode {
-    /// Normal mode
-    Normal,
-    PermissiveHold,
-    /// Trigger hold when any other non-tap-hold key is pressed
-    HoldOnOtherPress,
-}
-
-/// Configurations for tap hold behavior
-#[derive(Clone, Copy, Debug)]
-pub struct TapHoldConfig {
+pub struct MorseConfig {
     pub enable_hrm: bool,
     pub prior_idle_time: Duration,
-    pub hold_timeout: Duration,
-    pub mode: TapHoldMode,
+    /// Default timeout time for tap or hold
+    pub operation_timeout: Duration,
+    /// Default mode
+    pub mode: MorseKeyMode,
     /// If the previous key is on the same "hand", the current key will be determined as a tap
-    pub chordal_hold: bool,
+    pub unilateral_tap: bool,
 }
 
-impl Default for TapHoldConfig {
+impl Default for MorseConfig {
     fn default() -> Self {
         Self {
             enable_hrm: false,
-            chordal_hold: false,
-            mode: TapHoldMode::Normal,
+            unilateral_tap: false,
+            mode: MorseKeyMode::Normal,
             prior_idle_time: Duration::from_millis(120),
-            hold_timeout: Duration::from_millis(250),
+            operation_timeout: Duration::from_millis(250),
         }
     }
 }