+++ rmk/tests/common/mod.rs
@@ -1,3 +1,4 @@
+pub mod morse;
 pub mod test_macro;
 
 use core::cell::RefCell;
@@ -7,7 +8,7 @@ use embassy_futures::select::{Either, select};
 use embassy_sync::blocking_mutex::raw::CriticalSectionRawMutex;
 use embassy_sync::mutex::Mutex;
 use embassy_time::{Duration, Timer};
-use futures::{FutureExt, join};
+use futures::join;
 use log::debug;
 use rmk::action::KeyAction;
 use rmk::channel::{KEY_EVENT_CHANNEL, KEYBOARD_REPORT_CHANNEL};
@@ -59,21 +60,20 @@ pub async fn run_key_sequence_test<'a, const ROW: usize, const COL: usize, const
     join!(
         // Run keyboard until all reports are received
         async {
-            select(keyboard.run(), async {
-                select(
-                    Timer::after(MAX_TEST_TIMEOUT).then(|_| async {
-                        panic!("Test timeout reached");
-                    }),
-                    async {
-                        while !*REPORTS_DONE.lock().await {
-                            // polling reports
-                            Timer::after(Duration::from_millis(50)).await;
-                        }
-                    },
-                )
-                .await;
-            })
-            .await;
+            match select(
+                Timer::after(MAX_TEST_TIMEOUT),
+                select(keyboard.run(), async {
+                    while !*REPORTS_DONE.lock().await {
+                        // polling reports
+                        Timer::after(Duration::from_millis(50)).await;
+                    }
+                }),
+            )
+            .await
+            {
+                Either::First(_) => panic!("ERROR: report done timeout reached"),
+                _ => (),
+            }
         },
         // Send all key events with delays
         async {
@@ -89,37 +89,43 @@ pub async fn run_key_sequence_test<'a, const ROW: usize, const COL: usize, const
         },
         // Verify reports
         async {
-            let mut report_index = -1;
-            for expected in expected_reports {
-                match select(Timer::after(Duration::from_secs(2)), KEYBOARD_REPORT_CHANNEL.receive()).await {
-                    Either::First(_) => panic!("ERROR: report wait timeout reached"),
-                    Either::Second(Report::KeyboardReport(report)) => {
-                        report_index += 1;
-                        // println!("Received {}th report from channel: {:?}", report_index, report);
-                        assert_eq!(
-                            *expected, report,
-                            "on #{} reports, expected left but actually right",
-                            report_index
-                        );
-                    }
-                    Either::Second(report) => {
-                        debug!("other reports {:?}", report)
+            match select(Timer::after(MAX_TEST_TIMEOUT), async {
+                let mut report_index = -1;
+                for expected in expected_reports {
+                    match select(Timer::after(Duration::from_secs(2)), KEYBOARD_REPORT_CHANNEL.receive()).await {
+                        Either::First(_) => panic!("ERROR: report wait timeout reached"),
+                        Either::Second(Report::KeyboardReport(report)) => {
+                            report_index += 1;
+                            // println!("Received {}th report from channel: {:?}", report_index, report);
+                            assert_eq!(
+                                *expected, report,
+                                "on #{} reports, expected left but actually right",
+                                report_index
+                            );
+                        }
+                        Either::Second(report) => {
+                            debug!("other reports {:?}", report)
+                        }
                     }
                 }
-            }
 
-            // Wait for all key events to be sent
-            while !*SEQ_SEND_DONE.lock().await {
-                Timer::after(Duration::from_millis(50)).await;
-            }
+                // Wait for all key events to be sent
+                while !*SEQ_SEND_DONE.lock().await {
+                    Timer::after(Duration::from_millis(50)).await;
+                }
 
-            // Set done flag after all reports are verified
-            *REPORTS_DONE.lock().await = true;
+                // Set done flag after all reports are verified
+                *REPORTS_DONE.lock().await = true;
+            })
+            .await
+            {
+                Either::First(_) => panic!("Read report timeout"),
+                Either::Second(_) => (),
+            }
         }
     );
-    let buffer = keyboard.holding_buffer.clone();
-    if buffer.len() > 0 {
-        panic!("leak after buffer cleanup, buffer contains {:?}", buffer);
+    if !keyboard.held_buffer.is_empty() {
+        panic!("leak after buffer cleanup, buffer contains {:?}", keyboard.held_buffer);
     }
 }